% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Visualization.R
\name{pltDV}
\alias{pltDV}
\title{Plot data visualization for financial data sourced from yfinance}
\usage{
pltDV(
  symbol,
  from = "1000-05-01",
  to = "4000-01-01",
  SMA_period = 50,
  indicator = NULL
)
}
\arguments{
\item{symbol}{The symbol of data used for plotting (e.g., Stock Indices(AAPL, AMZN), Futures Contracts(CL=F, ),
Commodities(GOLD), Forex(EURUSD=X), etc).}

\item{from}{Start date for retrieving financial data (default: "1000-05-01").
If not specified, data will be retrieved from the earliest available date.}

\item{to}{End date for retrieving stock data (default: "4000-01-01").
If not specified, data will be retrieved up to the latest available date.}

\item{SMA_period}{Period for the simple moving average calculation (default: 50).}

\item{indicator}{Optional technical indicator to include in the visualization (choices: "RSI", "MACD", "Stochastic").}
}
\value{
A Plotly object displaying the stock visualization.
}
\description{
This function plots a visualization of financial data, including candlestick charts, simple moving averages,
volume bars, and optional technical indicators (e.g., RSI, MACD, Stochastic Oscillator).
}
\details{
If the \code{from} parameter is set to "1000-05-01" and the \code{to} parameter is set to "4000-01-01",
the function will retrieve the entire historical data available for the specified symbol.
This means it will fetch data from the earliest available date up to the latest available date.
}
\examples{
pltDV("AAPL", from = "2020-01-01", to = "2021-01-01",
SMA_period = 50, indicator = "RSI")

}
\seealso{
\code{\link{getSymbols}} function in the \code{quantmod} package.
}
